tests/wpt/tests/css/css-borders/tentative/parsing/border-shape-computed.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Borders and Box Decorations 4 Test: Parsing 'border-shape' with computed values</title>
<link rel="help" href="https://drafts.csswg.org/css-borders-4/#border-shape">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>

<div id="target"></div>

<script>
test_computed_value("border-shape", "none");
test_computed_value("border-shape", "circle()");
test_computed_value("border-shape", "polygon(10px 10px, 100px 10px, 10px 100px)");
test_computed_value("border-shape", "shape(from 0px 0px, hline to 100px, vline to 100px, close)");
test_computed_value("border-shape", "circle() circle()");
test_computed_value("border-shape", "circle() border-box circle() padding-box", "circle() circle()");
test_computed_value("border-shape", "circle() polygon(10px 10px, 100px 10px, 10px 100px)");
test_computed_value("border-shape", "circle() content-box");
test_computed_value("border-shape", "circle() border-box");
test_computed_value("border-shape", "circle() border-box polygon(10px 10px, 100px 10px, 10px 100px) border-box", "circle() polygon(10px 10px, 100px 10px, 10px 100px) border-box");
test_computed_value("border-shape", "circle() padding-box polygon(10px 10px, 100px 10px, 10px 100px) border-box", "circle() padding-box polygon(10px 10px, 100px 10px, 10px 100px) border-box");
test_computed_value("border-shape", "circle() padding-box polygon(10px 10px, 100px 10px, 10px 100px) padding-box", "circle() padding-box polygon(10px 10px, 100px 10px, 10px 100px)");
test_computed_value("border-shape", "circle() polygon(10px 10px, 100px 10px, 10px 100px) border-box", "circle() polygon(10px 10px, 100px 10px, 10px 100px) border-box");
test_computed_value("border-shape", "circle() polygon(10px 10px, 100px 10px, 10px 100px) content-box", "circle() polygon(10px 10px, 100px 10px, 10px 100px) content-box");
test_computed_value("border-shape", "circle() half-border-box", "circle()");
test_computed_value("border-shape", "circle() half-border-box circle() half-border-box", "circle()");
test_computed_value("border-shape", "circle() border-box circle() border-box", "circle() border-box");
</script>
